Add files via upload
A new project using an API in order to pull the price of certain crypto currencies, and show their value in specific standard currencies. Will be adding new functionality such as the ability to show price averages over set time spans, as well as adding some CSS styling to the page.